API_DEVELOPMENT/requirements.txt
pandas
numpy
geopy
matplotlib
seaborn
scikit-learn
tensorflow
joblib
requests
protobuf==3.20.*
fastapi
uvicorn[standard]
python-multipart
google-cloud-bigquery
pandas-gbq
db-dtypes